refactored login files in a folder for login
