+++ lib/features/auth/ui/login/widgets/do_not_have_account_widget.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
-import '../../../../core/constants/app_strings.dart';
-import '../../../../core/helpers/spaces.dart';
-import '../../../../core/themes/app_colors.dart';
-import '../../../../core/themes/app_text_styles.dart';
+import '../../../../../core/constants/app_strings.dart';
+import '../../../../../core/helpers/spaces.dart';
+import '../../../../../core/themes/app_colors.dart';
+import '../../../../../core/themes/app_text_styles.dart';
 
 class DoNotHaveAccountWidget extends StatelessWidget {
   const DoNotHaveAccountWidget({super.key});
@@ -19,8 +19,11 @@ class DoNotHaveAccountWidget extends StatelessWidget {
             color: AppColors.grey,
           ),
         ),
-        horizontalSpace(3),
+        horizontalSpace(5),
         TextButton(
+          style: TextButton.styleFrom(
+            padding: EdgeInsets.zero,
+          ),
           onPressed: () {},
           child: Text(
             AppStrings.register,
